Add basic auth in nginx
Add make target to generate the username:password file.

Although this file is stored in a directory called "public" that doesn't
mean it's public. This is just because we are using the static
buildpack, which uploads the contents of your directory in the "public"
directory. However, because we're using this buildpack as a
route-service, nginx is only used to manage requests and not to serve
files.